Move a utility function into allmydata.util.
